refactor: flattening the directory structure
