+++ internal/llmclient/client_test.go
