+++ internal/providers/anthropic/anthropic.go
@@ -16,7 +16,7 @@ import (
 	"github.com/google/uuid"
 
 	"gomodel/internal/core"
-	"gomodel/internal/pkg/llmclient"
+	"gomodel/internal/llmclient"
 	"gomodel/internal/providers"
 )
