+++ internal/providers/gemini/gemini.go
@@ -11,7 +11,7 @@ import (
 	"github.com/google/uuid"
 
 	"gomodel/internal/core"
-	"gomodel/internal/pkg/llmclient"
+	"gomodel/internal/llmclient"
 	"gomodel/internal/providers"
 )
